task_7_1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.0.1